+++ ebu_tt_live/scripts/ebu_simple_consumer.py
@@ -3,7 +3,8 @@ from argparse import ArgumentParser
 from .common import create_loggers
 
 from ebu_tt_live.node import SimpleConsumer
-from ebu_tt_live.twisted import TwistedConsumerImpl, TwistedConsumer, BroadcastClientFactory, ClientNodeProtocol
+from ebu_tt_live.twisted import TwistedConsumer, BroadcastClientFactory, ClientNodeProtocol
+from ebu_tt_live.carriage import TwistedConsumerImpl
 from twisted.internet import reactor
